[refactor] 휴대폰 번호 변경 컴포넌트 분리
